requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
matplotlib==3.6.3
numpy==1.24.1
